refactor(update): split integrity helpers
Move checksum asset lookup, checksum parsing, and file verification into a focused update_integrity module while re-exporting the existing simple_update API.

Tested: pytest tests/test_simple_update.py tests/test_split_update_timers.py tests/test_secure_temp_update.py tests/test_update_checksum_verification.py; ruff check src/accessiweather/services/simple_update.py src/accessiweather/services/update_integrity.py tests/test_simple_update.py tests/test_secure_temp_update.py tests/test_update_checksum_verification.py; ruff check src tests scripts; pyright.

Confidence: high
